requirements.txt
pyaudio
aubio
Pillow
numpy
PyQt5
# tkinter
